requirements.txt
Django==1.11.16
Pillow==5.3.0
boto3==1.9.87
botocore==1.12.87
certifi==2018.10.15
chardet==3.0.4
coverage==4.5.1
dj-database-url==0.5.0
django-forms-bootstrap==3.1.0
django-storages==1.7.1
django-vote==2.1.7
docutils==0.14
idna==2.7
jmespath==0.9.3
psycopg2==2.7.7
python-dateutil==2.8.0
pytz==2018.6
requests==2.20.1
s3transfer==0.1.13
stripe==2.12.0
urllib3==1.24.1
